requirements.txt
python-telegram-bot[job-queue]
requests
feedparser
yfinance
groq
python-dateutil